pyproject.toml
[project]
name = "mc-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.118.0",
    "greenlet>=3.2.4",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "resend>=2.16.0",
    "scalar-fastapi>=1.4.3",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "ruff>=0.13.3",
    "ty>=0.0.1a21",
]
